# execgen: remove COPYSLICE and WINDOW directives in favor of methods