--- a/pkg/col/coldata/vec_tmpl.go
+++ b/pkg/col/coldata/vec_tmpl.go
@@ -74,8 +74,8 @@ func (m *memColumn) Append(args SliceArgs) {
 				// bytes-like columns, we append an empty slice.
 				execgen.APPENDSLICE(toCol, toCol, args.DestIdx, 0, 0)
 				// {{else}}
-				// {{/* Here WINDOW means slicing which allows us to use APPENDVAL below. */}}
-				toCol = execgen.WINDOW(toCol, 0, args.DestIdx)
+				// {{/* Here Window means slicing which allows us to use APPENDVAL below. */}}
+				toCol = toCol.Window(0, args.DestIdx)
 				// {{end}}
 				for _, selIdx := range sel {
 					val := fromCol.Get(selIdx)
@@ -224,7 +224,7 @@ func (m *memColumn) Window(start int, end int) Vec {
 			return &memColumn{
 				t:                   m.t,
 				canonicalTypeFamily: m.canonicalTypeFamily,
-				col:                 execgen.WINDOW(col, start, end),
+				col:                 col.Window(start, end),
 				nulls:               m.nulls.Slice(start, end),
 			}
 			// {{end}}