--- a/pkg/sql/colexec/execgen/cmd/execgen/data_manipulation_gen.go
+++ b/pkg/sql/colexec/execgen/cmd/execgen/data_manipulation_gen.go
@@ -30,11 +30,6 @@ var dataManipulationReplacementInfos = []dataManipulationReplacementInfo{
 		replaceWith:         "CopyVal",
 	},
 	{
-		templatePlaceholder: "execgen.COPYSLICE",
-		numArgs:             5,
-		replaceWith:         "CopySlice",
-	},
-	{
 		templatePlaceholder: "execgen.APPENDSLICE",
 		numArgs:             5,
 		replaceWith:         "AppendSlice",