--- a/pkg/sql/colexec/execgen/cmd/execgen/data_manipulation_gen.go
+++ b/pkg/sql/colexec/execgen/cmd/execgen/data_manipulation_gen.go
@@ -40,11 +40,6 @@ var dataManipulationReplacementInfos = []dataManipulationReplacementInfo{
 		replaceWith:         "AppendVal",
 	},
 	{
-		templatePlaceholder: "execgen.WINDOW",
-		numArgs:             3,
-		replaceWith:         "Window",
-	},
-	{
 		templatePlaceholder: "execgen.SETVARIABLESIZE",
 		numArgs:             2,
 		replaceWith:         "SetVariableSize",