**coldata: ensure Len() methods are inlined**

Release note: None

**execgen: remove COPYSLICE directive in favor of the method**

This commit removes `execgen.COPYSLICE` directive in favor of
newly-introduced `CopySlice` method on the well-typed vectors. Most of
these methods are inlined (decimals are an exception in some cases).

Additionally, this commit removes some leftover stuff.

Release note: None

**execgen: remove WINDOW directive in favor of the method**

All calls of the newly-introduced method are actually inlined.

Release note: None